Fix clippy errors and warnings across all crates
- Fixed TopologyForecaster trait import in validator.rs and routing.rs
- Fixed thiserror String source field issues in mission/mod.rs and routing.rs
- Fixed InvalidIntent variant field errors in mission/compiler.rs
- Fixed ground_station_optical import errors in optical/tests/link_state_test.rs
- Fixed agent crate clippy warnings (unused imports, variables, Default implementations)
- Fixed borrow checker issue in resource/scheduler.rs
- Fixed unused variable warnings across oisl
- Fixed unused import warnings across oisl
- Fixed Duration type issue in control_node_assignment.rs
- Removed unused imports in tracking.rs

+++ acs/src/safety.rs
@@ -62,7 +62,7 @@ impl SafetyInterlock {
             el_min,
             el_max,
             emergency_stop: AtomicBool::new(false),
-            last_heartbeat: AtomicU64::new(Utc::now().timestamp_millis()),
+            last_heartbeat: AtomicU64::new(Utc::now().timestamp_millis() as u64),
             watchdog_timeout_ms,
             collision_zones: Vec::new(),
         }
@@ -117,7 +117,7 @@ impl SafetyInterlock {
     /// Update watchdog heartbeat
     pub fn update_watchdog(&self) {
         self.last_heartbeat
-            .store(Utc::now().timestamp_millis(), Ordering::Relaxed);
+            .store(Utc::now().timestamp_millis() as u64, Ordering::Relaxed);
     }
 
     /// Check watchdog timeout
